Seperated accordian toggle file from vehicle wrapper
