+++ client/src/components/VehicleDash/VehicleMainWrapper/index.js
@@ -1,48 +1,16 @@
 import React, { useEffect, useState, useContext } from "react";
 import "./style.css";
 import { useParams } from "react-router-dom";
+import ContextAwareToggle from "../../../utils/ContextAwareToggle"
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import Card from "react-bootstrap/Card";
 import Accordion from 'react-bootstrap/Accordion';
-import AccordionContext from 'react-bootstrap/AccordionContext';
-import { useAccordionToggle } from 'react-bootstrap/AccordionToggle';
 import openBtnIcon from "../../../images/vehiclepage/open-btn-icon.svg";
 import closeBtnIcon from "../../../images/vehiclepage/close-btn-icon.svg";
 import API from "../../../utils/API";
 import placeholder from "../../../images/user-page/car-type-icon.svg";
 import { useDriverKickContext } from "../../../utils/DriverKickContext";
-
-
-
-
-// START Bootstrap Accordion Component  ———————————————|
-function ContextAwareToggle({ children, eventKey, callback }) {
-  const currentEventKey = useContext(AccordionContext);
-  const { accordionHelper, setAccordionHelper } = useDriverKickContext();
-
-  const decoratedOnClick = useAccordionToggle(
-    eventKey,
-    () => {
-      callback && callback(eventKey)
-      accordionHelper ? setAccordionHelper(false) : setAccordionHelper(true)
-    }
-  );
-
-  const isCurrentEventKey = currentEventKey === eventKey;
-
-  return (
-    <button
-      type="button"
-      className="g__btn-reset"
-      onClick={decoratedOnClick}
-    >
-      {children}
-    </button>
-  );
-}
-// END Bootstrap Accordion Component  ———————————————|
-
 
 
 function VehicleMainWrapper({ children }) {
